+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## tip
 
+* FEATURE: Add links to VMUI from the query panel
+
 * BUGFIX: respect time filter for variables update. See [this issue](https://github.com/VictoriaMetrics/grafana-datasource/issues/47)
 
 ## [v0.1.2](https://github.com/VictoriaMetrics/grafana-datasource/releases/tag/v0.1.2)